Chore : 공통으로 사용되는 컴포넌트 그룹화
공통으로 사용되는 컴포넌트를 그룹으로 묶어 관리하였으며, 파일 경로를 수정하였습니다.
