+++ src/components/common/CustomSlider.jsx
